% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saviUserFunctions.R
\name{addRandomGroups}
\alias{addRandomGroups}
\title{Adds the vertex attribute "group" to each vertex in an igraph
randomly.}
\usage{
addRandomGroups(g, numOfVertexGroups)
}
\arguments{
\item{g}{The igraph to add vertex groups to.}

\item{numOfVertexGroups}{The desired number of vertex groups
for the graph.  It is required that this number be greater
than 1 or the function runSavi will throw an error.  It is
recommended that the number be less than the number of
verticies in the graph.}
}
\value{
The given igraph with each vertex in a random group.
}
\description{
Adds the vertex attribute "group" to each vertex in an igraph
randomly.
}
\examples{
    addRandomGroups(g,6)
}
