% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSaviHelperFunctions.R
\name{compareScore}
\alias{compareScore}
\title{Compare the scores of the current and proposed graph state.}
\usage{
compareScore(propScore, currScore, useStrict)
}
\arguments{
\item{propScore}{The score of the proposed graph state}

\item{currScore}{The score of the current graph state}

\item{useStrict}{A boolean.  If true, then the scores are
compared using strict less than.  If false, the scores are
compared using less than or equal to.  Note that the default
value passed by runSavi is FALSE.}
}
\value{
A the booelan result of the compared scores.
}
\description{
This is a helper function that is used in runSavi which is not
meant for the user.
}
\examples{
    compareScore(8,8,TRUE)
    compareScore(6,14,FALSE)
}
