% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saviUserFunctions.R
\name{createSaviGraph}
\alias{createSaviGraph}
\title{Creates a random igraph based on user specifications.}
\usage{
createSaviGraph(graphType, graphParams)
}
\arguments{
\item{graphType}{The graph type desired.  There are two types of
graphs currently available. The first is a random connected
graph ("Connected").  The other is the EFL graph ("EFL"),
which creates a random graph that conforms to the given
conditions of the Erdős–Faber–Lovász conjecture.}

\item{graphParams}{When the graph type is "Connected," the next
two arguements should be the number of verticies desired and
the second arguement should be the probability of each vertex
connecting.  Thus, this arguement should be in the interval
(0,1).  When the graph type is EFL, n is the number of
n-complete graphs to connect together.  A second argument
is not needed.}
}
\value{
The newly created igraph.
}
\description{
Creates a random igraph based on user specifications.
}
\examples{
    createSaviGraph("Connected",c(6,0.8))
    createSaviGraph("EFL",4)
}
