% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSaviHelperFunctions.R
\name{getPropGraphState}
\alias{getPropGraphState}
\title{Get a random proposed state of a given igraph.  To do this, the
function switches 2 vertices in different groups and returns an
igraph in the new state.}
\usage{
getPropGraphState(g, numOfVertexGroups)
}
\arguments{
\item{g}{The current igraph in its current state.}

\item{numOfVertexGroups}{The desired number of vertex groups for
the graph.  It is required that this number be greater than 1
or the function runSavi will throw an error.  It is recommended
that the number be less than the number of verticies in the
graph.}
}
\value{
A new igraph in the proposed state.
}
\description{
This is a helper function that is used in runSavi which is not
meant for the user.
}
\examples{
    getPropGraph(g,4)
}
