% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSaviHelperFunctions.R
\name{moveToPropGraphStateAnyway}
\alias{moveToPropGraphStateAnyway}
\title{Decide to move to the proposed graph state based on the current
and proposed graph states' scores.}
\usage{
moveToPropGraphStateAnyway(propScore, currScore, params)
}
\arguments{
\item{propScore}{The score of the proposed graph state.}

\item{currScore}{The score of the current graph state.}

\item{params}{A list of parameters based on the cooling schedule.}
}
\value{
boolean
}
\description{
This is a helper function that is used in runSavi which is not
meant for the user.
}
\examples{
    moveToPropGraphAnyway(10,12,params)
}
