% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saviUtilityFunctions.R
\name{sampleOne}
\alias{sampleOne}
\title{Return a random sample of size 1}
\usage{
sampleOne(x)
}
\arguments{
\item{x}{A vector of integers}
}
\value{
A random integer from x.  If x is c(5:5), return 5.
}
\description{
The default sample function is not compatible with vectors of
length 1.  If the code says "sample(3:3)," sample will return a
random integer between 1 and 3.  The way our code is written, we
would like it to return just 3.  We implement the following
function for that purpose.
}
\examples{
    sampleOne(c(4:10))
    sampleOne(c(4:4))

}
