% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSaviHelperFunctions.R
\name{updateParameters}
\alias{updateParameters}
\title{Update other parameters based on cooling schedule.}
\usage{
updateParameters(params, i)
}
\arguments{
\item{params}{The list of parameters}

\item{i}{The number of iterations of the loop}
}
\value{
params
}
\description{
This is a helper function that is used in runSavi.  It is not
meant for the user.
}
\examples{
    updateParameters(params,1,"Step")
}
